Finished day 1 part 2 in rust!
Why was there an offset of 2?

Take the string "7zmlcpsjneight7pbtqbkgl".

When searching for the last character, first the normal numbers where searched.
7 was then found at index 0, later 8 spelled out was searched and
because the index was later in the string the 7 was replaces by the
index of eight.

So because, after finding 7, there is no search for another 7 the last one
could not be found.

To fix this problem i changed from the &str.find function to &str.rfind
which instead of searching for the first match, searches for the last
match.

Another solution would have been to keep searching after a match. This
would however have resulted in more instructions for the program.